[ls-static-site] feat: move fn to own file
