+++ packages/larva/__tests__/lib/generateStatic.test.js
@@ -1,41 +1,11 @@
 const path = require( 'path' );
+const fs = require( 'fs' );
 const exec = require( 'child_process' ).exec;
 
-const fs = require( 'fs' );
-const mkdirp = require( 'mkdirp' );
-const http = require( 'http' );
-
 const fixture = path.join( __dirname, '../fixtures' );
+const generateStatic = require( '../../lib/generateStatic' );
 
 const buildPath = path.join( fixture, './build-html' );
-
-function generateStatic( routesArr, patternSource = 'larva' ) {
-
-	routesArr.map( ( route ) => {
-		const dir = path.join( buildPath, route );
-
-		mkdirp.sync( dir );
-
-		// TODO: handle async
-		const url = `http://localhost:3001/${patternSource}/${route}`;
-
-		http.get( url, res => {
-			let body = '';
-
-			res.setEncoding( 'utf8' );
-
-			res.on( 'data', data => {
-				body += data;
-			} );
-
-			res.on( 'end', () => {
-				fs.writeFileSync( `${dir}/index.html`, body );
-			} );
-		} );
-
-	} );
-
-}
 
 describe( 'generateStatic', () => {
 
@@ -56,7 +26,7 @@ describe( 'generateStatic', () => {
 			'components/c-link'
 		];
 
-		generateStatic( routesArr );
+		generateStatic( routesArr, buildPath );
 
 		expect(
 			fs.existsSync( path.join( buildPath, 'components/c-link/index.html' ) )
@@ -68,7 +38,7 @@ describe( 'generateStatic', () => {
 			'components/c-button/brand-basic'
 		];
 
-		generateStatic( routesArr );
+		generateStatic( routesArr, buildPath );
 
 		expect(
 			fs.existsSync( path.join( buildPath, 'components/c-button/brand-basic/index.html' ) )
@@ -82,7 +52,7 @@ describe( 'generateStatic', () => {
 			'modules/footer',
 		];
 
-		generateStatic( routesArr );
+		generateStatic( routesArr, buildPath );
 
 		expect(
 			fs.readFileSync(
